LayoutTests/fast/forms/textarea-appearance-wrap-expected.txt

This tests that textarea wrap attributes work correctly when submitting a form.






wrap=hard: Success


wrap=soft: Success


wrap=off: Success


wrap=hard rtl text: Success


wrap=hard mixed rtl and ltr text: Success


wrap=hard visibility:hidden: Success
